requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
librosa==0.10.1
torch==2.1.2
numpy==1.24.3
pydantic==2.5.3
python-multipart==0.0.6
soundfile==0.12.1
python-dotenv==1.0.0
scikit-learn==1.3.2
